Move XtApp wrappers into application.c
